Backend/requirements.txt
langchain_community
unstructured # Document loading
chromadb # Vector storage
langchain-google-genai # For embeddings
langchain_core
google-generativeai
python-dotenv
langchain-text-splitters
Flask